+++ backend/src/main/java/com/woowacourse/zzimkkong/controller/MemberController.java
@@ -12,8 +12,8 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import java.net.URI;
 
-import static com.woowacourse.zzimkkong.dto.Validator.EMAIL_MESSAGE;
-import static com.woowacourse.zzimkkong.dto.Validator.EMPTY_MESSAGE;
+import static com.woowacourse.zzimkkong.dto.ValidatorMessage.EMAIL_MESSAGE;
+import static com.woowacourse.zzimkkong.dto.ValidatorMessage.EMPTY_MESSAGE;
 
 @RestController
 @RequestMapping("/api/members")
