+++ backend/src/main/java/com/woowacourse/zzimkkong/dto/map/MapCreateUpdateRequest.java
@@ -2,7 +2,7 @@ package com.woowacourse.zzimkkong.dto.map;
 
 import javax.validation.constraints.NotBlank;
 
-import static com.woowacourse.zzimkkong.dto.Validator.EMPTY_MESSAGE;
+import static com.woowacourse.zzimkkong.dto.ValidatorMessage.EMPTY_MESSAGE;
 
 public class MapCreateUpdateRequest {
     @NotBlank(message = EMPTY_MESSAGE)
